+++ R/Module.R
@@ -103,10 +103,10 @@ Module <- R6::R6Class( # nolint object_name_linter
     print = function(style = TRUE) {
       registries <- self$get_registries()
 
-      is_link_table <- purrr::map(registries, "is_link_table") |>
+      is_link_table <- map(registries, "is_link_table") |>
         unlist()
 
-      standard_lines <- purrr::map_chr(
+      standard_lines <- map_chr(
         names(registries)[!is_link_table],
         function(.registry) {
           paste0("    $", registries[[.registry]]$class_name)
@@ -123,7 +123,7 @@ Module <- R6::R6Class( # nolint object_name_linter
         lines <- cli::ansi_strip(lines)
       }
 
-      purrr::walk(lines, cli::cat_line)
+      walk(lines, cli::cat_line)
     },
     #' @description
     #' Create a string representation of a `Module`
@@ -134,7 +134,7 @@ Module <- R6::R6Class( # nolint object_name_linter
     to_string = function(style = FALSE) {
       registries <- self$get_registries()
 
-      is_link_table <- purrr::map(registries, "is_link_table") |>
+      is_link_table <- map(registries, "is_link_table") |>
         unlist()
 
       registry_strings <- make_key_value_strings(
@@ -144,7 +144,7 @@ Module <- R6::R6Class( # nolint object_name_linter
             paste(
               paste0(
                 "$",
-                purrr::map_chr(registries[!is_link_table], "class_name")
+                map_chr(registries[!is_link_table], "class_name")
               ),
               collapse = ", "
             ),