+++ R/laminr-package.R
@@ -5,6 +5,6 @@
 #' @importFrom cli cli_abort cli_warn cli_inform
 #' @importFrom R6 R6Class
 #' @importFrom httr GET POST content add_headers
-#' @importFrom purrr map map_chr map2 pmap set_names list_flatten transpose discard keep
+#' @importFrom purrr map map_chr map_lgl map2 pmap set_names list_flatten transpose discard keep list_c walk reduce
 ## usethis namespace: end
 NULL