+++ CHANGELOG.md
@@ -46,6 +46,8 @@ For more information, please visit the [package website](https://laminr.lamin.ai
 
 * Adjust colours in print output (PR #69)
 
+* Modify `Registry` print output to separate relational fields by module (PR #71)
+
 ## TESTING
 
 - Add a simple unit test which queries laminlabs/lamindata (PR #27).